backend/requirements.txt
black
python-dotenv
psycopg2
flask
pytest